% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fit.R
\name{extract_fit}
\alias{extract_fit}
\title{Extract SEM model information}
\usage{
extract_fit(text)
}
\arguments{
\item{text}{A character vector (paragraph) copied from academic paper}
}
\value{
A list with elements:
\itemize{
\item model_type: character vector of detected model type keywords (SEM/MSEM/DSEM)
\item fit_indices: character vector of detected fit indices (CFI/TLI/RMSEA/SRMR)
\item description: character vector of qualitative fit adjective
}
}
\description{
Extract SEM model information
}
\examples{
finnigan <- paste(
"We conducted multi-level structural equation modeling (ML-SEM)",
"with the R package lavaan."
)
extract_fit(finnigan)
}
